Revert "Changed Product id type from Long to Integer (due to sql issue)"
This reverts commit a92c4d2b98b4f9c4384457ffaab5e2c1a4bb290a.

+++ app/controllers/HomeController.java
@@ -82,7 +82,7 @@ public class HomeController extends Controller {
     }
 
     // Delete Product by id
-    public Result deleteProduct(Integer id) {
+    public Result deleteProduct(Long id) {
 
         // find product by id and call delete method
         Product.find.ref(id).delete();
